+++ src/test/java/io/github/astrapi69/time/convert/OffsetDateTimeExtensionsTest.java
@@ -28,36 +28,39 @@ import de.alpharogroup.date.CreateDateExtensions;
 import org.junit.jupiter.api.Test;
 import org.meanbean.test.BeanTester;
 
+import java.sql.Timestamp;
 import java.time.LocalDateTime;
+import java.time.OffsetDateTime;
 import java.time.ZoneId;
+import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
 import java.util.Date;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
- * The unit test class for the class {@link ZonedDateTimeExtensions}
+ * The unit test class for the class {@link OffsetDateTimeExtensions}
  */
 public class OffsetDateTimeExtensionsTest
 {
 
 	/**
-	 * Test method for {@link ZonedDateTimeExtensions#toDate(ZonedDateTime)}
+	 * Test method for {@link OffsetDateTimeExtensions#toDate(OffsetDateTime)}
 	 */
 	@Test
 	void testToDate()
 	{
 		Date actual;
 		Date expected;
-		ZonedDateTime input;
+		OffsetDateTime input;
 		ZoneId zoneId;
 		LocalDateTime localDateTime;
 
 		localDateTime = LocalDateTime.of(2000, 9, 1, 0, 0, 0);
 		zoneId = ZoneId.systemDefault();
 
-		input = ZonedDateTime.of(localDateTime, zoneId);
-		actual = ZonedDateTimeExtensions.toDate(input);
+		input = localDateTime.atZone(zoneId).toOffsetDateTime();
+		actual = OffsetDateTimeExtensions.toDate(input);
 		expected = CreateDateExtensions.newDate(2000, 9, 1);
 		if (!actual.equals(expected))
 		{
@@ -70,7 +73,7 @@ public class OffsetDateTimeExtensionsTest
 
 
 	/**
-	 * Test method for {@link ZonedDateTimeExtensions#toZonedDateTime(LocalDateTime, ZoneId)}
+	 * Test method for {@link OffsetDateTimeExtensions#toZonedDateTime(OffsetDateTime)}
 	 */
 	@Test
 	void testToZonedDateTime()
@@ -79,32 +82,94 @@ public class OffsetDateTimeExtensionsTest
 		ZonedDateTime expected;
 		ZoneId zoneId;
 		LocalDateTime localDateTime;
+		OffsetDateTime offsetDateTime;
 
 		localDateTime = LocalDateTime.of(2000, 9, 1, 0, 0, 0);
 		zoneId = ZoneId.of("Europe/Paris");
-
-		actual = ZonedDateTimeExtensions.toZonedDateTime(localDateTime, zoneId);
-		expected = ZonedDateTime.parse("2000-09-01T00:00+02:00[Europe/Paris]");
+		offsetDateTime = OffsetDateTimeExtensions.toOffsetDateTime(localDateTime, zoneId);
+		actual = OffsetDateTimeExtensions.toZonedDateTime(offsetDateTime);
+		expected = ZonedDateTime.parse("2000-09-01T00:00+02:00");
 		assertEquals(expected, actual);
 	}
 
 	/**
-	 * Test method for {@link ZonedDateTimeExtensions#toZonedDateTime(Date, ZoneId)}
+	 * Test method for {@link OffsetDateTimeExtensions#toOffsetDateTime(Date, ZoneId)
 	 */
 	@Test
-	void testToZonedDateTimeFromDateAndZoneId()
+	void testToOffsetDateTime()
 	{
 
-		ZonedDateTime actual;
-		ZonedDateTime expected;
+		OffsetDateTime actual;
+		OffsetDateTime expected;
 		ZoneId zoneId;
 		Date date;
 
 		date = CreateDateExtensions.newDate(2000, 9, 1, 0, 0, 0);
 		zoneId = ZoneId.of("Europe/Paris");
 
-		actual = ZonedDateTimeExtensions.toZonedDateTime(date, zoneId);
-		expected = ZonedDateTime.parse("2000-09-01T00:00+02:00");
+		actual = OffsetDateTimeExtensions.toOffsetDateTime(date, zoneId);
+		expected = OffsetDateTime.parse("2000-09-01T00:00+02:00");
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link OffsetDateTimeExtensions#toOffsetDateTime(LocalDateTime, ZoneId)}
+	 */
+	@Test
+	void testToOffsetDateTimeWithLocalDateTimeZoneId()
+	{
+
+		OffsetDateTime actual;
+		OffsetDateTime expected;
+		ZoneId zoneId;
+		LocalDateTime localDateTime;
+
+		localDateTime = LocalDateTime.of(2000, 9, 1, 0, 0, 0);
+		zoneId = ZoneId.of("Europe/Paris");
+
+		actual = OffsetDateTimeExtensions.toOffsetDateTime(localDateTime, zoneId);
+		expected = OffsetDateTime.parse("2000-09-01T00:00+02:00");
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link OffsetDateTimeExtensions#toOffsetDateTime(LocalDateTime, ZoneOffset)}
+	 */
+	@Test
+	void testToOffsetDateTimeWithLocalDateTimeZoneOffset()
+	{
+
+		OffsetDateTime actual;
+		OffsetDateTime expected;
+		ZoneOffset zoneOffSet;
+		LocalDateTime localDateTime;
+
+		localDateTime = LocalDateTime.of(2000, 9, 1, 0, 0, 0);
+		zoneOffSet = ZoneOffset.of("+02:00");
+
+		actual = OffsetDateTimeExtensions.toOffsetDateTime(localDateTime, zoneOffSet);
+		expected = OffsetDateTime.parse("2000-09-01T00:00+02:00");
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link OffsetDateTimeExtensions#toTimestamp(OffsetDateTime)}
+	 */
+	@Test
+	void testToTimestamp()
+	{
+		Timestamp actual;
+		Timestamp expected;
+		ZoneId zoneId;
+		LocalDateTime localDateTime;
+		OffsetDateTime offsetDateTime;
+
+		localDateTime = LocalDateTime.of(2000, 9, 1, 0, 0, 0);
+		zoneId = ZoneId.of("Europe/Paris");
+		offsetDateTime = OffsetDateTimeExtensions.toOffsetDateTime(localDateTime, zoneId);
+
+		actual = OffsetDateTimeExtensions.toTimestamp(offsetDateTime);
+		expected = LocalDateTimeExtensions.toTimestamp(localDateTime);
 		assertEquals(expected, actual);
 	}