+++ CHANGELOG.md
@@ -7,6 +7,14 @@ Version 2.3-SNAPSHOT
 ADDED:
 
 - new method for convert LocalDate to legacy Date
+- new method for convert LocalDateTime with a ZoneId to an OffsetDateTime
+- new method for convert LocalDateTime with a ZoneOffset to an OffsetDateTime
+- new method for convert OffsetDateTime to a Timestamp
+- new method for convert LocalDateTime to a Timestamp
+
+CHANGED:
+
+- update of gradle-plugin dependency 'com.diffplug.spotless:spotless-plugin-gradle' in version 6.11.0
 
 Version 2.2
 -------------